requirements.txt
Flask
flask_cors
bcrypt
pyjwt
python-dotenv
pytz
openpyxl
pyodbc
pymssql
